Se agrego clases repository y service

+++ src/main/java/com/bda/TrabajoPractico/Entidades/Tarifa.java
@@ -19,7 +19,7 @@ public class Tarifa {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private double id;
+	private Integer id;
 	
 	@Column(name="TIPO_TARIFA")
 	private int tipoTarifa;
